Refactor `search` module to be part of `beatmapsets`
