+++ app/routes/public/beatmap.py
@@ -1,8 +1,7 @@
 
 from app.common.database import beatmaps, scores, favourites, nominations, relationships
 from app.common.constants import BeatmapLanguage, BeatmapGenre, DatabaseStatus, Mods
-
-from flask import Blueprint, request
+from flask import Blueprint, request, redirect
 from flask_login import current_user
 
 import config
@@ -11,7 +10,11 @@ import app
 
 router = Blueprint('beatmap', __name__)
 
-@router.get('/<id>')
+@router.get('/beatmaps/<id>')
+def redirect_to_map(id: int):
+    return redirect(f'/b/{id}')
+
+@router.get('/b/<id>')
 def get_beatmap(id: int):
     if not id.isdigit():
         return utils.render_error(404, 'beatmap_not_found')
